+++ src/main/java/racing_game/view/RacingGameInput.java
@@ -3,13 +3,14 @@ package racing_game.view;
 import java.util.Scanner;
 
 public class RacingGameInput {
-  private static Scanner scanner = new Scanner(System.in);
+
+  private static final Scanner scanner = new Scanner(System.in);
 
   /**
    * 자동차 이름 입력 메세지 출력 후 스캐너 호출
-   * 
+   * <p>
    * 이름은 쉼표(,)로 구분한다.
-   * 
+   *
    * @return String[], 자동차 이름 목록 배열
    */
   public static String[] getNameOfCarsByScanner() {
@@ -20,7 +21,7 @@ public class RacingGameInput {
 
   /**
    * 횟수 입력 메세지 출력 후 스캐너 호출
-   * 
+   *
    * @return int, 반복 횟수
    */
   public static int getTimeByScanner() {
