+++ src/main/java/racing_game/view/RacingGameResultView.java
@@ -1,9 +1,10 @@
 package racing_game.view;
 
 import java.util.List;
-import racing_game.model.Car;
+import racing_game.model.domain.Car;
 
 public class RacingGameResultView {
+
   private final static String RESULT_BEGIN_MESSAGE = "\n실행결과";
   private final static String SYMBOL_DISTANCE = "-";
   private final static String RESULT_WINNER_MESSAGE = "(이)가 최종우승했습니다.";
@@ -17,10 +18,10 @@ public class RacingGameResultView {
 
   /**
    * 입력 받은 이름과 거리에 따라 대쉬(-) 출력 <br/>
-   * 
+   * <p>
    * e.g. name : -
-   * 
-   * @param name : String, 출력할 이름
+   *
+   * @param name     : String, 출력할 이름
    * @param distance : int 출력할 거리
    */
   public static void printNameAndDistance(String name, int distance) {
@@ -35,7 +36,7 @@ public class RacingGameResultView {
 
   /**
    * 자동차들의 이름과 거리 출력
-   * 
+   *
    * @param racingCarList : 출력할 자동차 목록
    */
   public static void printNameAndDistancesOfRacingCarList(List<Car> racingCarList) {
@@ -48,14 +49,14 @@ public class RacingGameResultView {
 
   /**
    * 자동차 경주의 우승자 출력
-   * 
-   * @param racingCarList : 출력할 자동차 목록
+   *
+   * @param winnerList : 출력할 자동차 목록
    */
   public static void printWinner(List<Car> winnerList) {
-    StringBuffer message = new StringBuffer();
+    StringBuilder message = new StringBuilder();
 
     for (Car winner : winnerList) {
-      message.append(winner.getName() + ", ");
+      message.append(winner.getName()).append(", ");
     }
 
     message.delete(message.length() - 2, message.length());
