+++ src/test/java/racing_game/view/RacingGameInputTest.java
@@ -1,13 +1,12 @@
 package racing_game.view;
 
-//import static org.assertj.core.api.Assertions.*;
-
 import org.junit.jupiter.api.Test;
 
 class RacingGameInputTest {
 
   @Test
   void getNameOfCarsByScanner() {
+    RacingGameInput.getNameOfCarsByScanner();
   }
 
   @Test
